pyproject.toml
[project]
name = "make-planning-notes-for-nate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ollama>=0.4.7",
    "pygithub>=2.5.0",
    "python-dotenv>=1.0.1",
    "pytz>=2025.1",
]
